GuiScreens are weird
This commit fixes a bug where the "cancel" state of the generateImagesButton would seemingly not do anything if the left side of it was clicked. Why did this happen?

Here is how buttons are updated in a GuiScreen:

 - A for loop is ran on the size of the button list in the Gui (checked every iteration)
 - A button is retrieved by then calling the "get" method by the index of the counter variable
 - The button then tests for being clicked and drawing etc

The reason this code wan't working previously was because I called initGui during the for loop. Although it didn't break anything per say, it did mean that the for loop got to run an extra button - it had popped into existence during the loop, but that was "fine", because for every iteration, it would re-check how big the list of buttons was. The new button happened to be shifted over by one position - the previewRecipeImagesButton got added, but the index was pointing to the button after it, which was now the generateImagesButton. This meant that the generateImagesButton would effectively just get fired twice in rapid succession, provided you clicked in the left area of it (where the new button was). The solution is just to always have the same amount of buttons in the Gui, and to disable one if it's not in use. It's probably for the best.